+++ nglab/__init__.py
@@ -1,0 +1,5 @@
+"""
+nglab - Next Gen Laboratory
+
+Multimodal Deep Reinforcement Learning for financial trading.
+"""